Start fixing lint error
