+++ Gemfile
@@ -1,1 +1,3 @@
+source 'https://rubygems.org'
+
 gem 'rubocop', '>= 1.0', '< 2.0'
